requirements.txt
requests==2.27.1
loguru==0.5.3
Django==3.2.6
PyJWT==2.4.0
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
Pillow==9.2.0
